// releases/20260221_131156/src/payload-types.ts
export interface ContentBlock {
  richText?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'content';
}
// ...
export interface Template {
  id: number;
  title: string;
  content?: ContentBlock[] | null;
  updatedAt: string;
  createdAt: string;
}